Install gulp globally and locally.
First we verify if node and npm are installed. Now we can use `npm init` and
initialize the project with some defaults. This creates the `package.json`.

Now we will install gulp with `npm install -g gulp` to install the **global**
gulp command-line-interface (CLI). With that installed we can then `npm install
gulp --save-dev` to install the local version of gulp for the project.

With gulp installed, we can now create the `gulpfile.js` that will contain our
tasks. In the gulpfile, we will require gulp itself and create a `default` task
that simply says hello.